+++ Domain/Customer.php
@@ -17,71 +17,11 @@ namespace Bookstore\Domain;
  * 
  */
   
-// abstract classes cannot be instantiated
-abstract class Customer extends Person {
-
-    // making static property
-    private static $lastId = 0;
-
-    private $id;
-    // private $firstname;
-    // private $surname;
-    private $email;
-
-    public function __construct($id, $firstname, $surname, $email){
-        // $this->id = $id;
-
-        parent::__construct($firstname, $surname);
-
-        if ($id == null){
-            // accessing static property
-            $this->id = ++self::$lastId;
-        }else {
-            $this->id == $id;
-            if($id> self::$lastId){
-                self::$lastId = $id;
-            }
-        }
-
-        $this->firstname = $firstname;
-        $this->surname = $surname;
-        $this->email = $email;
-    }
+// interface can extend from another interface
+Interface Customer extends Payer{
 
     // abstract methods
-    abstract public function getMonthlyFee();
-    abstract public function getAmountToBorrow();
-    abstract public function getType();
-
-    // accessor 
-    public static function getLastId(){
-        return self::$lastId;
-    }
-
-    // using getter to get name
-    public function getFirstName(){
-        return $this->firstname;
-    }
-
-    public function getSurname(){
-    return $this->surname;
-    }
-
-    public function getFullName(){
-        return $this->firstname . ' ' . $this->surname;
-    }
-
-    public function getEmail(){
-        return $this->email;
-    }
-
-    // using setter/mutators to set email
-    public function setEmail(string $email){
-        $this->email = $email;
-    }
-
-    public function sayHi(){
-        // getting both child and parent implementation
-        return 'Howdy ' . $this->firstname . parent::sayHi();
-    }
+    public function getMonthlyFee();
+    public function getAmountToBorrow();
+    public function getType();
 }
