+++ Domain/Customer/Basic.php
@@ -3,9 +3,10 @@
 namespace Bookstore\Domain\Customer;
 
 use Bookstore\Domain\Customer;
+use Bookstore\Domain\Person;
 
 
-class Basic extends Customer {
+class Basic extends Person implements Customer {
 
     public function getMonthlyFee(){
         return 5.0;
@@ -17,5 +18,13 @@ class Basic extends Customer {
 
     public function getType() {
         return 'Basic';
+    }
+
+    public function pay(float $amount){
+        return "Paying $amount";
+    }
+
+    public function isExtentOfTaxes(){
+        return "You gotta pay tax mate";
     }
 }
