+++ Domain/Payer.php
@@ -1,0 +1,8 @@
+<? php 
+
+namespace Bookstore\Domain;
+
+interface Payer {
+    public function pay(float $amont);
+    public function isExtentOfTaxes()
+}
